Move Query to query package
