+++ server/storage-jdbc/src/main/java/org/bithon/server/storage/jdbc/metric/MetricJdbcReader.java
@@ -28,8 +28,8 @@ import org.bithon.server.storage.jdbc.dsl.sql.StringExpression;
 import org.bithon.server.storage.jdbc.utils.SQLFilterBuilder;
 import org.bithon.server.storage.metrics.IFilter;
 import org.bithon.server.storage.metrics.IMetricReader;
-import org.bithon.server.storage.metrics.OrderBy;
-import org.bithon.server.storage.metrics.Query;
+import org.bithon.server.storage.datasource.query.OrderBy;
+import org.bithon.server.storage.datasource.query.Query;
 import org.jooq.DSLContext;
 import org.jooq.Field;
 import org.jooq.Record;
