+++ server/storage/src/main/java/org/bithon/server/storage/datasource/query/OrderBy.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.metrics;
+package org.bithon.server.storage.datasource.query;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
