+++ server/storage/src/main/java/org/bithon/server/storage/metrics/IMetricReader.java
@@ -18,6 +18,7 @@ package org.bithon.server.storage.metrics;
 
 import org.bithon.server.commons.time.TimeSpan;
 import org.bithon.server.storage.datasource.DataSourceSchema;
+import org.bithon.server.storage.datasource.query.Query;
 
 import java.util.Collection;
 import java.util.List;
