+++ server/web-service/src/main/java/org/bithon/server/web/service/datasource/api/GeneralQueryRequest.java
@@ -18,9 +18,9 @@ package org.bithon.server.web.service.datasource.api;
 
 import lombok.Data;
 import org.bithon.server.storage.metrics.IFilter;
-import org.bithon.server.storage.metrics.Limit;
-import org.bithon.server.storage.metrics.OrderBy;
-import org.bithon.server.storage.metrics.Query;
+import org.bithon.server.storage.datasource.query.Limit;
+import org.bithon.server.storage.datasource.query.OrderBy;
+import org.bithon.server.storage.datasource.query.Query;
 
 import javax.annotation.Nullable;
 import javax.validation.Valid;
