+++ server/web-service/src/main/java/org/bithon/server/web/service/datasource/api/ListQueryRequest.java
@@ -18,7 +18,7 @@ package org.bithon.server.web.service.datasource.api;
 
 import lombok.Data;
 import org.bithon.server.storage.metrics.IFilter;
-import org.bithon.server.storage.metrics.OrderBy;
+import org.bithon.server.storage.datasource.query.OrderBy;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
