+++ server/web-service/src/main/java/org/bithon/server/web/service/datasource/api/impl/DataSourceApi.java
@@ -31,9 +31,9 @@ import org.bithon.server.storage.metrics.IMetricReader;
 import org.bithon.server.storage.metrics.IMetricStorage;
 import org.bithon.server.storage.metrics.IMetricWriter;
 import org.bithon.server.storage.metrics.Interval;
-import org.bithon.server.storage.metrics.Limit;
+import org.bithon.server.storage.datasource.query.Limit;
 import org.bithon.server.storage.metrics.MetricStorageConfig;
-import org.bithon.server.storage.metrics.Query;
+import org.bithon.server.storage.datasource.query.Query;
 import org.bithon.server.web.service.datasource.api.DataSourceService;
 import org.bithon.server.web.service.datasource.api.DisplayableText;
 import org.bithon.server.web.service.datasource.api.GeneralQueryRequest;
